requirements.txt
evdev==1.2.0
pyusb==1.0.2
PyYAML==5.2
